pyproject.toml
[project]
name = "MedContext-impact"
version = "0.1.0"
description = "MedContext-impact leverages MedGemma to tackle the emerging issue of medical deepfakes."
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Jamie Forrest" }]
license = { text = "MIT" }
dependencies = [
  "fastapi==0.115.5",
  "uvicorn==0.30.6",
  "pydantic==2.9.2",
  "pydantic-settings==2.5.2",
  "SQLAlchemy==2.0.35",
  "python-multipart==0.0.12",
  "alembic==1.13.3",
  "httpx==0.27.2",
  "google-cloud-aiplatform==1.71.0",
  "langgraph==0.2.40",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"